+++ src/main/java/be/cytomine/repository/project/ProjectRepository.java
@@ -21,12 +21,11 @@ import be.cytomine.domain.project.Project;
 import be.cytomine.domain.security.User;
 import be.cytomine.dto.DatedCytomineDomain;
 import be.cytomine.dto.NamedCytomineDomain;
+import jakarta.persistence.Tuple;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import jakarta.persistence.Tuple;
-import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
@@ -58,7 +57,7 @@ public interface ProjectRepository extends JpaRepository<Project, Long>  {
     List<Tuple> listByCreator(Long userId);
 
     default List<NamedCytomineDomain> listByCreator(User user) {
-        return listByCreator(user.getId()).stream().map(x -> new NamedCytomineDomain(((BigInteger)x.get(0)).longValue(), (String)x.get(1)))
+        return listByCreator(user.getId()).stream().map(x -> new NamedCytomineDomain((Long) x.get(0), (String)x.get(1)))
                 .collect(Collectors.toList());
     }
 
@@ -66,7 +65,7 @@ public interface ProjectRepository extends JpaRepository<Project, Long>  {
     List<Tuple> listByAdmin(Long userId);
 
     default List<NamedCytomineDomain> listByAdmin(User user) {
-        return listByAdmin(user.getId()).stream().map(x -> new NamedCytomineDomain(((BigInteger)x.get(0)).longValue(), (String)x.get(1)))
+        return listByAdmin(user.getId()).stream().map(x -> new NamedCytomineDomain((Long) x.get(0), (String)x.get(1)))
                 .collect(Collectors.toList());
     }
 
@@ -74,7 +73,7 @@ public interface ProjectRepository extends JpaRepository<Project, Long>  {
     List<Tuple> listByUser(Long userId);
 
     default List<NamedCytomineDomain> listByUser(User user) {
-        return listByUser(user.getId()).stream().map(x -> new NamedCytomineDomain(((BigInteger)x.get(0)).longValue(), (String)x.get(1)))
+        return listByUser(user.getId()).stream().map(x -> new NamedCytomineDomain((Long) x.get(0), (String)x.get(1)))
                 .collect(Collectors.toList());
     }
 
@@ -86,12 +85,12 @@ public interface ProjectRepository extends JpaRepository<Project, Long>  {
     List<Tuple> listLastCreatedTuple();
 
     default List<DatedCytomineDomain> listLastCreated(List<Long> ignoredProjectIds) {
-        return listLastCreatedTuple(ignoredProjectIds).stream().map(x -> new DatedCytomineDomain(((BigInteger)x.get(0)).longValue(), (Date)x.get(1)))
+        return listLastCreatedTuple(ignoredProjectIds).stream().map(x -> new DatedCytomineDomain((Long) x.get(0), (Date)x.get(1)))
                 .collect(Collectors.toList());
     }
 
     default List<DatedCytomineDomain> listLastCreated() {
-        return listLastCreatedTuple().stream().map(x -> new DatedCytomineDomain(((BigInteger)x.get(0)).longValue(), (Date)x.get(1)))
+        return listLastCreatedTuple().stream().map(x -> new DatedCytomineDomain((Long) x.get(0), (Date)x.get(1)))
                 .collect(Collectors.toList());
     }
