+++ src/main/java/be/cytomine/service/utils/TaskService.java
@@ -24,13 +24,11 @@ import be.cytomine.service.database.SequenceService;
 import be.cytomine.service.security.SecurityACLService;
 import be.cytomine.utils.Task;
 import be.cytomine.utils.TaskComment;
+import jakarta.persistence.EntityManager;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import jakarta.persistence.EntityManager;
-import jakarta.transaction.Transactional;
-
-import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -189,10 +187,10 @@ public class TaskService {
         }
 
         Task task = new Task();
-        task.setId(((BigInteger)row[0]).longValue());
-        task.setProgress(((BigInteger)row[1]).intValue());
-        task.setProjectIdent(row[2]!=null ? ((BigInteger)row[2]).longValue() : null);
-        task.setUserIdent(((BigInteger)row[3]).longValue());
+        task.setId((Long) row[0]);
+        task.setProgress((Integer) row[1]);
+        task.setProjectIdent(row[2]!=null ? (Long) row[2] : null);
+        task.setUserIdent((Long) row[3]);
         return task;
     }
 
@@ -203,7 +201,7 @@ public class TaskService {
         return rows.stream().map(x -> {
             TaskComment comment = new TaskComment();
             comment.setComment((String)x[0]);
-            comment.setTimestamp(((BigInteger)x[1]).longValue());
+            comment.setTimestamp((Long) x[1]);
             return comment;
         }).collect(Collectors.toList());
     }
